+++ packages/react-native-kit/src/components/atoms/Avatar/Avatar.tsx
@@ -2,6 +2,8 @@ import { Avatar as AvatarCore, AvatarProps } from '@tecsinapse/react-core';
 import React from 'react';
 import { Text } from '../Text';
 
-export const Avatar: React.FC<AvatarProps> = props => {
+const Avatar: React.FC<AvatarProps> = props => {
   return <AvatarCore {...props} TextComponent={Text} />;
 };
+
+export default Avatar;