+++ packages/react-native-kit/src/components/atoms/Tag/index.ts
@@ -1,1 +1,2 @@
-export { default as Tag, TagProps } from './Tag';
+export { default as Tag } from './Tag';
+export type { TagProps } from './Tag';