+++ packages/react-native-kit/src/components/atoms/Text/index.ts
@@ -1,1 +1,2 @@
-export { default as Text, TextNativeProps } from './Text';
+export { default as Text } from './Text';
+export type { TextNativeProps } from './Text';