+++ packages/react-native-kit/src/components/atoms/TextArea/index.ts
@@ -1,1 +1,2 @@
-export { default as TextArea, TextAreaProps } from './TextArea';
+export { default as TextArea } from './TextArea';
+export type { TextAreaProps } from './TextArea';