+++ packages/react-native-kit/src/components/molecules/DatePicker/index.ts
@@ -1,1 +1,2 @@
-export { DatePicker, NativeDatePickerProps } from './DatePicker';
+export { DatePicker } from './DatePicker';
+export type { NativeDatePickerProps } from './DatePicker';