+++ packages/react-web-kit/src/components/atoms/Table/Footer/TFoot.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { TFooter } from './styled';
 
-export const TFoot: React.FC<React.HTMLAttributes<HTMLTableSectionElement>> = ({
+const TFoot: React.FC<React.HTMLAttributes<HTMLTableSectionElement>> = ({
   children,
   ...rest
 }) => {