+++ packages/react-web-kit/src/components/atoms/Table/Header/Th.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { ThStyled } from './styled';
 
-export const Th: React.FC<
-  React.ThHTMLAttributes<HTMLTableHeaderCellElement>
-> = ({ children, ...rest }) => {
+const Th: React.FC<React.ThHTMLAttributes<HTMLTableHeaderCellElement>> = ({
+  children,
+  ...rest
+}) => {
   return <ThStyled {...rest}>{children}</ThStyled>;
 };